Fix: LandingPage route defect
