+++ src/pages/IndexPage.vue
@@ -1,53 +1,6 @@
 <template>
   <q-page class="relative-position overflow-hidden" style="min-height: 100vh">
     <!-- ═══════════════════════════════════════════════ -->
-    <!-- HERO WELCOME OVERLAY (Landing Page)            -->
-    <!-- ═══════════════════════════════════════════════ -->
-    <transition name="hero-fade">
-      <div v-if="showWelcomeOverlay" class="absolute-full z-max hero-overlay">
-        <q-img
-          src="https://phworldexpo.tpb.gov.ph/wp-content/uploads/2025/05/Lake-Lanao.png"
-          class="absolute-full"
-          fit="cover"
-          style="filter: brightness(0.4)"
-        />
-
-        <div class="particles-container absolute-full">
-          <div v-for="n in 20" :key="n" class="particle" :style="particleStyle(n)" />
-        </div>
-
-        <div class="absolute-full flex flex-center">
-          <div class="text-center q-pa-xl hero-content" style="position: relative; z-index: 5">
-            <div class="q-mb-lg">
-              <q-avatar size="90px" class="hero-logo-avatar">
-                <q-icon name="water_drop" size="50px" color="white" />
-              </q-avatar>
-            </div>
-
-            <h1 class="hero-title q-mb-none">RANAO FISHNET</h1>
-            <div class="hero-divider q-mx-auto q-my-md" />
-            <p class="hero-subtitle q-mb-xs">Profiling and Mapping of Lake Lanao Fishes</p>
-            <p class="hero-description q-mb-xl">
-              A Web-Based GIS Platform for Ecological Research and Conservation
-            </p>
-
-            <q-btn unelevated rounded size="lg" class="explore-btn q-px-xl" @click="enterDashboard">
-              <q-icon name="travel_explore" class="q-mr-sm" size="sm" />
-              Explore the Map
-            </q-btn>
-
-            <div class="row q-gutter-xl justify-center q-mt-xl">
-              <div class="hero-stat" v-for="stat in heroStats" :key="stat.label">
-                <div class="hero-stat-value">{{ stat.value }}</div>
-                <div class="hero-stat-label">{{ stat.label }}</div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </transition>
-
-    <!-- ═══════════════════════════════════════════════ -->
     <!-- MAIN DASHBOARD (Map + Control Panel)           -->
     <!-- ═══════════════════════════════════════════════ -->
 
@@ -55,9 +8,8 @@
     <div ref="mapContainer" class="absolute-full map-container" />
 
     <!-- Toggle Button -->
-    <transition name="fade-btn">
+    <transition name="fade-btn" appear>
       <q-btn
-        v-if="!showWelcomeOverlay"
         class="toggle-panel-btn"
         :class="{ 'toggle-btn--shifted': showPanel }"
         round
@@ -74,7 +26,7 @@
 
     <!-- Floating Control Panel (Left Side — Toggleable, BRIGHT THEME) -->
     <transition name="slide-panel">
-      <div v-if="showPanel && !showWelcomeOverlay" class="control-panel">
+      <div v-if="showPanel" class="control-panel">
         <q-card class="bright-panel full-height column no-wrap">
           <!-- Panel Header -->
           <q-card-section class="q-pb-xs">
@@ -448,7 +400,7 @@
 
     <!-- ═══ DETAIL POPUP (Right Side — BRIGHT THEME) ═══ -->
     <transition name="slide-detail">
-      <div v-if="(selectedFish || selectedWaterSite) && !showWelcomeOverlay" class="detail-panel">
+      <div v-if="selectedFish || selectedWaterSite" class="detail-panel">
         <q-card class="bright-panel full-height column">
           <q-card-section class="q-pb-sm">
             <div class="row items-center justify-between">
@@ -604,7 +556,7 @@
     <!-- ═══ ADD DATA DROPDOWN BUTTON (for logged-in researchers) ═══ -->
     <transition name="fade-btn">
       <q-btn
-        v-if="!showWelcomeOverlay && authStore.isLoggedIn"
+        v-if="authStore.isLoggedIn"
         class="add-data-btn"
         :class="{ 'add-data-btn--shifted': showPanel }"
         unelevated
@@ -672,19 +624,17 @@
 
 <script setup lang="ts">
 import { ref, computed, onMounted, watch, nextTick } from 'vue';
-import { useRoute, useRouter } from 'vue-router';
+import { useRouter } from 'vue-router';
 import { useQuasar } from 'quasar';
 import { useAuthStore } from 'src/stores/auth';
 import L from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 
-const route = useRoute();
 const router = useRouter();
 const authStore = useAuthStore();
 const $q = useQuasar();
 
 // ═══ STATE ═══
-const showWelcomeOverlay = ref(true);
 const activeTab = ref('fish');
 const fishSearch = ref('');
 const activeFilter = ref('all');
@@ -706,14 +656,6 @@ let currentBaseTileLayer: L.TileLayer | null = null;
 // Lake Lanao boundary rings ([lat, lng][]) — populated once the boundary GeoJSON
 // loads, used to detect "click anywhere inside the lake" for the reading popup.
 let lakePolygonRings: [number, number][][] = [];
-
-// ═══ HERO STATS ═══
-const heroStats = [
-  { value: '24', label: 'Species Recorded' },
-  { value: '18', label: 'Endemic Cyprinids' },
-  { value: '6', label: 'Invasive Species' },
-  { value: '12', label: 'Critically Endangered' },
-];
 
 // ═══ FISH DATA ═══
 interface Fish {
@@ -1505,18 +1447,6 @@ function closeDetailPanel() {
   selectedWaterSite.value = null;
 }
 
-function particleStyle(n: number) {
-  const size = 2 + Math.random() * 4;
-  return {
-    width: `${size}px`,
-    height: `${size}px`,
-    left: `${Math.random() * 100}%`,
-    top: `${Math.random() * 100}%`,
-    animationDelay: `${Math.random() * 8}s`,
-    animationDuration: `${6 + Math.random() * 10}s`,
-  };
-}
-
 // ═══ MAP INITIALIZATION ═══
 function initMap() {
   if (!mapContainer.value || map) return;
@@ -1719,21 +1649,13 @@ watch(filteredSpecies, () => {
   renderFishMarkers();
 });
 
-function enterDashboard() {
-  showWelcomeOverlay.value = false;
+onMounted(() => {
   nextTick(() => {
     initMap();
     setTimeout(() => {
       map?.invalidateSize();
     }, 400);
   });
-}
-
-onMounted(() => {
-  // If navigated with ?explore=true, skip the welcome overlay and go straight to map
-  if (route.query.explore === 'true') {
-    enterDashboard();
-  }
 });
 
 // Watch layer toggle changes and sync map visibility
@@ -1760,103 +1682,6 @@ function goToWaterQuality() {
 </script>
 
 <style scoped>
-/* ═══════════════════════════════════ */
-/* HERO OVERLAY                       */
-/* ═══════════════════════════════════ */
-.hero-overlay {
-  z-index: 9999;
-}
-
-.hero-title {
-  font-size: 3.5rem;
-  font-weight: 900;
-  color: white;
-  letter-spacing: 6px;
-  text-shadow: 0 4px 20px rgba(0, 0, 0, 0.8);
-}
-
-.hero-divider {
-  width: 80px;
-  height: 3px;
-  background: linear-gradient(90deg, transparent, #26a69a, transparent);
-}
-
-.hero-subtitle {
-  font-size: 1.3rem;
-  font-weight: 400;
-  color: #e0e0e0;
-  text-shadow: 0 2px 10px rgba(0, 0, 0, 0.6);
-}
-
-.hero-description {
-  font-size: 0.95rem;
-  color: #9e9e9e;
-  text-shadow: 0 2px 8px rgba(0, 0, 0, 0.5);
-}
-
-.hero-logo-avatar {
-  background: linear-gradient(135deg, #00695c, #26a69a) !important;
-  box-shadow: 0 4px 30px rgba(38, 166, 154, 0.4);
-}
-
-.explore-btn {
-  background: linear-gradient(135deg, #00695c 0%, #26a69a 100%) !important;
-  color: white;
-  font-weight: 700;
-  letter-spacing: 1px;
-  font-size: 1rem;
-  box-shadow: 0 4px 25px rgba(38, 166, 154, 0.5);
-  transition: all 0.3s ease;
-}
-.explore-btn:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 8px 35px rgba(38, 166, 154, 0.7);
-}
-
-.hero-stat {
-  text-align: center;
-}
-.hero-stat-value {
-  font-size: 1.6rem;
-  font-weight: 800;
-  color: #26a69a;
-  text-shadow: 0 0 20px rgba(38, 166, 154, 0.3);
-}
-.hero-stat-label {
-  font-size: 0.7rem;
-  color: #9e9e9e;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-}
-
-/* Floating Particles */
-.particles-container {
-  pointer-events: none;
-  z-index: 1;
-}
-.particle {
-  position: absolute;
-  background: rgba(38, 166, 154, 0.3);
-  border-radius: 50%;
-  animation: float-particle linear infinite;
-}
-@keyframes float-particle {
-  0% {
-    transform: translateY(0) scale(1);
-    opacity: 0;
-  }
-  20% {
-    opacity: 1;
-  }
-  80% {
-    opacity: 1;
-  }
-  100% {
-    transform: translateY(-100vh) scale(0.3);
-    opacity: 0;
-  }
-}
-
 /* ═══════════════════════════════════ */
 /* MAP                                */
 /* ═══════════════════════════════════ */
@@ -1999,15 +1824,6 @@ function goToWaterQuality() {
 /* ═══════════════════════════════════ */
 /* TRANSITIONS                        */
 /* ═══════════════════════════════════ */
-.hero-fade-enter-active,
-.hero-fade-leave-active {
-  transition: opacity 1s ease;
-}
-.hero-fade-enter-from,
-.hero-fade-leave-to {
-  opacity: 0;
-}
-
 .slide-panel-enter-active {
   transition: all 0.6s cubic-bezier(0.16, 1, 0.3, 1);
 }
