+++ src/pages/auth/LoginPage.vue
@@ -166,7 +166,7 @@ function handleLogin() {
     authStore.login(username.value, password.value);
     adminStore.logActivity(authStore.displayName, 'Logged In', 'Signed in to Ranao FishNet');
     loading.value = false;
-    router.push(authStore.user?.role === 'Admin' ? '/admin' : '/');
+    router.push(authStore.user?.role === 'Admin' ? '/admin' : '/map');
   }, 800);
 }
